placement interference: fix erroneous detection after long invisible ball
and a ball-misdetection.

The check is not executed as long as the ball is invisible and therefore
the times for the robots are not updated to World.Time (the changed
line).
If the ball is invisible for more than 2 seconds and then appears near a
non-placing teams robot, it would immediately call for placement
interference (wether the ball was really there or not), not giving the
robot the required 2 seconds to drive away.

Setting it to nil instead solves the issue as the inRangeStartTime will
be updated when the ball is actually visible and too close to the robot.