replay test runner: add option to ignore some troublesome tests
Some tests might not be feasible to be fixed since they are related to
the tracking or other complex reasons.
Rather than having the pipeline continously fail, it is better to ignore
these cases and open an issue for them.
This is also important since other teams can also add test cases.