Goal: evaluate closeToGoal with respect to goalline
Previously closeToGoal was evaluated by comparing the distance of the Ball
Since then the goal size has changed and the magic constant is no longer
a good value.

In principle, it is better to calculate the distance to the goalline instead
of the center of the goal in the first place.
Also the magic constant is at least declared as a global variable at the top
of the file for better readability.